Refactoring, like button is now component
