+++ src/views/HomeView.vue
@@ -1,3 +1,97 @@
 <template>
-    <h1 class="text-3xl font-bold">BooksAPI</h1>
-</template>
+    <h1 class="text-3xl font-bold underline">BooksAPI</h1>
+    <div class="relative z-10" aria-labelledby="modal-title" role="dialog" aria-modal="true" v-if="selectedBook">
+    <div class="fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity"></div>
+  
+    <div class="fixed inset-0 z-10 w-screen overflow-y-auto">
+      <div class="flex min-h-full items-end justify-center p-4 text-center sm:items-center sm:p-0">
+        <div class="relative transform overflow-hidden rounded-lg bg-white text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-lg">
+          <div class="bg-white px-4 pb-4 pt-5 sm:p-6 sm:pb-4">
+            <div class="sm:flex sm:items-start">
+              <div class="mx-auto flex h-12 w-12 flex-shrink-0 items-center justify-center rounded-full bg-red-100 sm:mx-0 sm:h-10 sm:w-10">
+                <svg class="h-6 w-6 text-red-600" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" aria-hidden="true">
+                  <path stroke-linecap="round" stroke-linejoin="round" d="M12 9v3.75m-9.303 3.376c-.866 1.5.217 3.374 1.948 3.374h14.71c1.73 0 2.813-1.874 1.948-3.374L13.949 3.378c-.866-1.5-3.032-1.5-3.898 0L2.697 16.126zM12 15.75h.007v.008H12v-.008z" />
+                </svg>
+              </div>
+              <div class="mt-3 text-center sm:ml-4 sm:mt-0 sm:text-left">
+                <h3 class="text-base font-semibold leading-6 text-gray-900" id="modal-title">{{ selectedBook.title }}</h3>
+                <div class="mt-2">
+                  <p class="text-sm text-gray-500">{{ selectedBook.description }}</p>
+                </div>
+              </div>
+            </div>
+          </div>
+          <div class="bg-gray-50 px-4 py-3 sm:flex sm:flex-row-reverse sm:px-6">
+            <LikeButton :isLiked="selectedBook.isLiked" @like="likeAction(selectedBook)"/>
+            <button @click="closeBookDetails" type="button" class="mt-3 inline-flex w-full justify-center rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50 sm:mt-0 sm:w-auto">Cancel</button>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+  
+    <ul role="list" class="divide-y divide-gray-100 grid grid-cols-3 mx-20">
+      <li
+        v-for="book in books"
+        :key="book.id"
+        class="flex justify-between gap-x-6 py-5 mx-8"
+        @click="openBookDetails(book)"
+      >
+        <div class="flex min-w-0 gap-x-4">
+          <img
+            class="h-16 w-16 flex-none rounded-full bg-gray-50"
+            :src="book.cover_image"
+            alt=""
+          />
+          <div class="min-w-0 flex-auto">
+            <p class="text-sm font-semibold leading-6 text-gray-900">
+              {{ book.title }}
+            </p>
+            <p class="mt-1 truncate text-xs leading-5 text-gray-500">
+              {{ book.author }}
+            </p>
+          </div>
+        </div>
+        <div class="hidden shrink-0 sm:flex sm:flex-col sm:items-end">
+          <p class="text-sm leading-6 text-gray-900">
+            {{ book.publication_year }}
+          </p>
+          <LikeButton :isLiked="book.isLiked" @like="likeAction(book)"/>
+        </div>
+      </li>
+    </ul>
+  </template>
+  
+  <script setup>
+  import { ref, onMounted } from "vue";
+  import axios from "axios";
+  
+  import LikeButton from '../components/LikeButton.vue'
+
+  const books = ref([]);
+  const selectedBook = ref(null)
+  
+  const getBooks = async () => {
+    try {
+      const response = await axios.get("https://freetestapi.com/api/v1/books");
+      books.value = response.data;
+    } catch (error) {
+      console.error("Error fetching books:", error);
+    }
+  };
+  
+  onMounted(getBooks);
+  
+  const likeAction = (book) => {
+    book.isLiked = !book.isLiked;
+  };
+  
+  const openBookDetails = (book) => {
+    selectedBook.value = { ...book };
+  };
+  
+  const closeBookDetails = () => {
+    selectedBook.value = null;
+  };
+  
+  </script>
